whatpm/t/webidl/webidl-interface.dat

#data
interface abc;
#errors
1;non-module definition;s
#document
| interface abc;

#data
interface abc : def;
#errors
1.20;before interface block;m
#document

#data
interface abc : def;
interface xyz;
#errors
1.20;before interface block;m
2;non-module definition;s
#document
| interface xyz;

#data
module m {
interface x {
  attribute long a;
  attribute long b;
};};
#errors
#document
| module m {
| 
| interface x {
|   attribute long a;
|   attribute long b;
| };
| 
| };

#data
module m {
interface x {
  attribute long a;
  attribute long a;
};};
#errors
4;duplicate member;m
#document
| module m {
| 
| interface x {
|   attribute long a;
|   attribute long a;
| };
| 
| };

#data
module m {
interface x {
  attribute long a;
};
interface x {
  attribute long a;
};};
#errors
5;duplicate qname;m

#data
module m {
interface x {
  attribute long a;
  attribute long _a;
};};
#errors
4;duplicate member;m

#data
module m {
interface x {
  attribute long a;
  attribute long __a;
};};
#errors

#data
module m {
interface x {
  attribute long a;
  const long a = 1;
};};
#errors
4;duplicate member;m

#data
module m {
  interface x {
    attribute long a;
  };
  interface y {
    attribute long a;
  };
};
#errors

#data
module m {
interface x {
  long a ();
  long a ();
};};
#errors

#data
interface x {};
interface y : x {};
#errors
1;non-module definition;s
2;non-module definition;s

#data
interface x;
interface y : x {};
#errors
1;non-module definition;s
2;non-module definition;s

#data
interface x : y {};
#errors
1;non-module definition;s
1;interface not defined;m;y

#data
module m {
  interface x {};
  interface y : x {};
};
#errors

#data
module m {
  interface x {};
};
module n {
  interface y : x {};
};
#errors
5;interface not defined;m;x

#data
module m {
  interface x : y::DOMString {};
};
#errors
2;interface not defined;m;y::DOMString

#data
module m {
  interface _DOMString {};
  interface x : _DOMString {};
};
#errors

#data
module m {
  interface _DOMString {};
  interface x : m::_DOMString {};
};
#errors
3;interface not defined;m;m::DOMString

#data
module m {
  interface _DOMString {};
  interface x : m::DOMString {};
};
#errors
3;interface not defined;m;m::DOMString

#data
module m {
  interface _DOMString {};
  interface x : ::m::_DOMString {};
};
#errors
3;interface not defined;m;::m::DOMString

#data
interface x {}; // defines x1
module m {
  interface y : x {}; // inherits x1
  interface x : x {}; // defines x2, which inherits x1
  interface z : x {}; // inherits x2
};
#errors
1;non-module definition;s

#data
module m {
  interface x : x {};
};
#errors
2;interface not defined;m;x

#data
module m {
  interface y : x {};
  interface x {};
};
#errors
2;interface not defined;m;x

#data
module m {
  interface x {};
  interface y {};
  interface z : x, y {};
};
#errors

#data
module m {
  interface x : ::x {};
};
#errors
2;interface not defined;m;::x

#data
module m {
  interface x {};
};
module n {
  interface y : m::x {};
};
#errors

#data
module m {
  interface x {};
  module n {
    interface y : x {};
  };
};
#errors
4;interface not defined;m;x

#data
module m {
  module n {
    interface x {};
  };
};
module l {
  interface y : m::n::x {};
};
#errors

#data
module m {
  module n {};
  interface x : n {};
};
#errors
3;interface not defined;m;n

#data
module m {
  exception x {};
  interface y : x {};
};
#errors
3;interface not defined;m;x

#data
module m {
  interface x;
  interface y : x {};
};
#errors

#data
module m {
  interface i {
    const long t = 234;
  };
};
#errors

#data
module m {
  interface i {
    const boolean t = 234;
  };
};
#errors
3:234;const type mismatch;m;boolean

#data
module m {
  interface i {
    const unsigned long t = -02;
  };
};
#errors
3:-2;const value out of range;m

#data
module m {
  interface i {
    const any t = 234;
  };
};
#errors

#data
module m {
  interface i {
    const long t = 234.4;
  };
};
#errors
3:234.4;const type mismatch;m;long

#data
module m {
  interface i {
    const long t =TRUE;
  };
};
#errors
3:TRUE;const type mismatch;m;long

#data
module m {
  interface i {
    const DOMString t = 234;
  };
};
#errors
3:234;const type mismatch;m;DOMString

#data
module m {
  const octet LF = 0x0A;
};
#errors
#document
| module m {
| 
| const octet LF = 10;
| 
| };

#data
module m {
  const octet LF = 0X0a;
};
#errors
#document
| module m {
| 
| const octet LF = 10;
| 
| };

#data
module m {
  const octet LF = 012;
};
#errors
#document
| module m {
| 
| const octet LF = 10;
| 
| };

#data
module m {
  const octet LF = 10;
};
#errors
#document
| module m {
| 
| const octet LF = 10;
| 
| };

#data
module m {
  const octet LF = 10.00;
};
#errors
2:10;const type mismatch;m;octet
#document
| module m {
| 
| const octet LF = 10;
| 
| };

#data
module m {
  const octet CR = -0x0D;
};
#errors
2:-13;const value out of range;m
#document
| module m {
| 
| const octet CR = -13;
| 
| };

#data
interface x {};
module m {
  typedef long x;
  interface y : x {};
};
#errors
1;non-module definition;s
4;interface not defined;m;x

#data
module m {
  interface x {
    attribute long a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute long a;
| };
| 
| };

#data
module m {
  interface x {
    attribute unsigned short a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute unsigned short a;
| };
| 
| };

#data
module m {
  interface x {
    attribute Object a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute Object a;
| };
| 
| };

#data
module m {
  interface x {
    attribute DOMString a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute DOMString a;
| };
| 
| };

#data
module m {
  interface x {
    attribute _DOMString a;
  };
};
#errors
3;type not defined;m;DOMString

#data
module m {
  interface x {
    attribute m::DOMString a;
  };
};
#errors
3;type not defined;m;m::DOMString
#document
| module m {
| 
| interface x {
|   attribute m::DOMString a;
| };
| 
| };

#data
module m {
  interface x {
    attribute m::_DOMString a;
  };
};
#errors
3;type not defined;m;m::DOMString

#data
module m {
  exception ex {};
  interface x {
    attribute ex x;
  };
};
#errors
4;type not defined;m;ex

#data
module m {
  typedef long l;
  interface x {
    attribute l mylong;
  };
};
#errors

#data
module m {
  valuetype mystr sequence<octet>;
  interface x {
    attribute mystr text;
  };
};
#errors

#data
module m {
  const boolean v = TRUE;
  interface x {
    attribute v myvalue;
  };
};
#errors
4;type not defined;m;v

#data
module m {
  interface x {
    attribute x a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute x a;
| };
| 
| };

#data
module m {
  interface x {
    attribute m::x a;
  };
};
#errors
#document
| module m {
| 
| interface x {
|   attribute m::x a;
| };
| 
| };

#data
module m {
  interface y {};
  interface x {
    attribute y a;
  };
};
#errors
#document
| module m {
| 
| interface y {
| };
| interface x {
|   attribute y a;
| };
| 
| };

#data
module m {
  interface x {
    attribute y a;
  };
  interface y {
  };
};
#errors
3;type not defined;m;y
#document
| module m {
| 
| interface x {
|   attribute y a;
| };
| interface y {
| };
| 
| };

#data
interface y {};
module m {
  interface x {
    attribute y a;
  };
};
#errors
1;non-module definition;s
#document
| interface y {
| };
| 
| module m {
| 
| interface x {
|   attribute y a;
| };
| 
| };

#data
interface y {};
module m {
  interface x {
    attribute ::y a;
  };
};
#errors
1;non-module definition;s
#document
| interface y {
| };
| 
| module m {
| 
| interface x {
|   attribute ::y a;
| };
| 
| };

#data
module m {
  interface y {};
  interface x {
    attribute ::m::y a;
  };
};
#errors
#document
| module m {
| 
| interface y {
| };
| interface x {
|   attribute ::m::y a;
| };
| 
| };

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x);
  };
};
#errors
#document
| module m {
| 
| exception x {
| };
| interface y {
|   attribute boolean a getraises (x);
| };
| 
| };

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x,x);
  };
};
#errors
#document
| module m {
| 
| exception x {
| };
| interface y {
|   attribute boolean a getraises (x, x);
| };
| 
| };

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x, ::m::x);
  };
};
#errors
#document
| module m {
| 
| exception x {
| };
| interface y {
|   attribute boolean a getraises (x, ::m::x);
| };
| 
| };

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x, z);
  };
};
#errors
4;exception not defined;m;z

#data
module m {
  
  interface y {
    attribute boolean a getraises (z);
  };
};
#errors
4;exception not defined;m;z

#data
module m {
  exception _DOMString {};
  interface y {
    attribute boolean a getraises (_DOMString);
  };
};
#errors

#data
module m {
  exception _DOMString {};
  interface y {
    attribute boolean a getraises (m::_DOMString);
  };
};
#errors
4;exception not defined;m;m::DOMString

#data
exception _DOMString {};
module m {
  interface y {
    attribute boolean a getraises (::_DOMString);
  };
};
#errors
1;non-module definition;s
4;exception not defined;m;::DOMString

#data
module m {
  interface y {
    attribute boolean a getraises (x);
  };
  exception x {};
};
#errors
3;exception not defined;m;x

#data
module m {
  interface x {};
  interface y {
    attribute boolean a getraises (x);
  };
};
#errors
4;exception not defined;m;x

#data
module m {
  typedef long x;
  interface y {
    attribute boolean a getraises (x);
  };
};
#errors
4;exception not defined;m;x

#data
module m {
  exception ex {}; typedef ex x;
  interface y {
    attribute boolean a getraises (x);
  };
};
#errors
4;exception not defined;m;x

#data
module m {
  valuetype x boolean;
  interface y {
    attribute boolean a getraises (x);
  };
};
#errors
4;exception not defined;m;x

#data
module m {
  exception x {};
  interface y {
    attribute boolean a setraises (x);
  };
};
#errors

#data
module m {
 
  interface y {
    attribute boolean a setraises (x);
  };
};
#errors
4;exception not defined;m;x

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x) setraises (x);
  };
};
#errors

#data
module m {
  exception x {};
  interface y {
    attribute boolean a getraises (x) setraises (z);
  };
};
#errors
4;exception not defined;m;z

#data
module m {
  interface x {
    void m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   void m ();
| };
| 
| };

#data
module m {
  interface x {
    long m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   long m ();
| };
| 
| };

#data
module m {
  interface x {
    long long m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   long long m ();
| };
| 
| };

#data
module m {
  interface x {
    DOMString m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   DOMString m ();
| };
| 
| };

#data
module m {
  interface x {
    x m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   x m ();
| };
| 
| };

#data
module m {
  interface x {
    m::x m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   m::x m ();
| };
| 
| };

#data
module m {
  interface x {
    ::m::x m ();
  };
};
#errors
#document
| module m {
| 
| interface x {
|   ::m::x m ();
| };
| 
| };

#data
module m {
  interface x {
    m method ();
  };
};
#errors
3;type not defined;m;m

#data
module m {
  interface x {
    _DOMString method ();
  };
};
#errors
3;type not defined;m;DOMString

#data
module m {
  interface _DOMString {};
  interface x {
    _DOMString method ();
  };
};
#errors

#data
module m {
  interface _DOMString {};
  interface x {
    m::_DOMString method ();
  };
};
#errors
4;type not defined;m;m::DOMString

#data
module m {
  interface u {};
  interface x {
    u method ();
  };
};
#errors

#data
interface f {};
module m {
  interface x {
    f method ();
  };
};
#errors
1;non-module definition;s

#data
module m {
  interface x {
    z method ();
  };
};
#errors
3;type not defined;m;z

#data
module m {
  typedef long z;
  interface x {
    z method ();
  };
};
#errors

#data
module m {
  interface x {
    _long method ();
  };
};
#errors
3;type not defined;m;long

#data
module m {
  valuetype s sequence<DOMString>;
  interface x {
    s method ();
  };
};
#errors

#data
module m {
  exception e {};
  interface x {
    e method ();
  };
};
#errors
4;type not defined;m;e

#data
module m {
  interface x {
    void method (in boolean b);
  };
};
#errors
#document
| module m {
| 
| interface x {
|   void method (in boolean b);
| };
| 
| };

#data
module m {
  interface x {
    void method (in boolean b, in long c);
  };
};
#errors
#document
| module m {
| 
| interface x {
|   void method (in boolean b, in long c);
| };
| 
| };

#data
module m {
  interface x {
    void method (in boolean _b, in long _c);
  };
};
#errors
#document
| module m {
| 
| interface x {
|   void method (in boolean b, in long c);
| };
| 
| };

#data
module m {
  interface x {
    void method (in DOMString b);
  };
};
#errors
#document
| module m {
| 
| interface x {
|   void method (in DOMString b);
| };
| 
| };

#data
module m {
  interface x {
    void method (in x b);
  };
};
#errors

#data
module m {
  interface x {
    void method (in _x b);
  };
};
#errors

#data
module m {
  interface x {
    void method (in m::x b);
  };
};
#errors

#data
module m {
  interface x {
    void method (in ::m::x b);
  };
};
#errors

#data
module m {
  typedef sequence<DOMString> ss;
  interface x {
    void method (in ss b);
  };
};
#errors

#data
module m 
{
  valuetype a long;
  interface x {
    void method (in a b);
  };
};
#errors

#data
module m {
  exception ex {};
  interface x {
    void method (in ex b);
  };
};
#errors
4;type not defined;m;ex

#data
module m {
  interface x {
    void method (in _DOMString b);
  };
};
#errors
3;type not defined;m;DOMString

#data
module m {
  interface x {
    void method (in y b);
  };
  interface y {};
};
#errors
3;type not defined;m;y

#data
module m {
  exception e {};
  interface x {
    void method () raises (e);
  };
};
#errors
#document
| module m {
| 
| exception e {
| };
| interface x {
|   void method () raises (e);
| };
| 
| };

#data
module m {
  exception e {};
  interface x {
    void method () raises (e,e);
  };
};
#errors
#document
| module m {
| 
| exception e {
| };
| interface x {
|   void method () raises (e, e);
| };
| 
| };

#data
module m {
  exception e {};
  exception f {};
  interface x {
    void method () raises (e, f);
  };
};
#errors
#document
| module m {
| 
| exception e {
| };
| exception f {
| };
| interface x {
|   void method () raises (e, f);
| };
| 
| };

#data
module m {
  interface x {
    void method () raises (e);
  };
};
#errors
3;exception not defined;m;e

#data
module m {
  interface x {
    void method () raises (e);
  };
  exception e {};
};
#errors
3;exception not defined;m;e

#data
module m {
  interface e {};
  interface x {
    void method () raises (e);
  };
};
#errors
4;exception not defined;m;e

#data
module m {
  valuetype e long;
  interface x {
    void method () raises (e);
  };
};
#errors
4;exception not defined;m;e

#data
module m {
  typedef DOMString e;
  interface x {
    void method () raises (e);
  };
};
#errors
2;typedef ignored;i
4;exception not defined;m;e

#data
module m {
  exception ee {};
  valuetype e ee;
  interface x {
    void method () raises (e);
  };
};
#errors
3;not boxable type;m;ee
5;exception not defined;m;e

#data
module m {
  [Constructor] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor] interface x {
| };
| 
| };

#data
module m {
  [constructor] interface x {
  };
};
#errors
2;unknown xattr;u;constructor
#document
| module m {
| 
| [constructor] interface x {
| };
| 
| };

#data
module m {
  [Constructor()] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor ()] interface x {
| };
| 
| };

#data
module m {
  [Constructor(in boolean b)] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (in boolean b)] interface x {
| };
| 
| };

#data
module m {
  [Constructor (in long c, in DOMString g)] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (in long c, in DOMString g)] interface x {
| };
| 
| };

#data
module m {
  [Constructor( [u] in short  h)] interface x {
  };
};
#errors
2;unknown xattr;u;u
#document
| module m {
| 
| [Constructor ([u] in short h)] interface x {
| };
| 
| };

#data
module m {
  [Constructor  (in b  h, in x i)] interface x {
  };
};
#errors
2;type not defined;m;b
#document
| module m {
| 
| [Constructor (in b h, in x i)] interface x {
| };
| 
| };

#data
module m {
  [Constructor  (),Constructor] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (), Constructor] interface x {
| };
| 
| };

#data
module m {
  [Constructor  (),Constructor (in boolean c)] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (), Constructor (in boolean c)] interface x {
| };
| 
| };

#data
module m {
  [Constructor  (in boolean c),Constructor (in boolean c)] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (in boolean c), Constructor (in boolean c)] interface x {
| };
| 
| };

#data
module m {
  [_Constructor] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor] interface x {
| };
| 
| };

#data
module m {
  [_Constructor (in boolean _a, in _x b)] interface x {
  };
};
#errors
#document
| module m {
| 
| [Constructor (in boolean a, in x b)] interface x {
| };
| 
| };

#data
module m {
  [Constructor=a] interface x {
  };
};
#errors
2;xattr id not allowed;m;Constructor
#document
| module m {
| 
| [Constructor=a] interface x {
| };
| 
| };

#data
module m {
  [Constructor = b ( )] interface x {
  };
};
#errors
2;xattr id not allowed;m;Constructor
#document
| module m {
| 
| [Constructor=b ()] interface x {
| };
| 
| };

#data
module m {
  [Constructor=b (in long c)] interface x {
  };
};
#errors
2;xattr id not allowed;m;Constructor
#document
| module m {
| 
| [Constructor=b (in long c)] interface x {
| };
| 
| };

#data
module m {
  [Constructor=x([u] in long c)] interface x {
  };
};
#errors
2;xattr id not allowed;m;Constructor
2;unknown xattr;u;u
#document
| module m {
| 
| [Constructor=x ([u] in long c)] interface x {
| };
| 
| };

#data
module m {
  interface x {
    [Constructor] void m ();
  };
};
#errors
3;xattr not applicable;w;Constructor

#data
[Constructor] module m {
  interface x {
    void m ();
  };
};
#errors
1;xattr not applicable;w;Constructor

#data
module m {
  interface x {
    [Constructor] attribute long c;
  };
};
#errors
3;xattr not applicable;w;Constructor

#data
module m {
  [Constructor] interface x;
  interface x { };
};
#errors
2;xattr not applicable;w;Constructor
3;duplicate qname;m

#data
module m {
  interface x;
  [Constructor] interface x { };
};
#errors
3;duplicate qname;m

#data
module m {
  [Constructor] interface x;
  [Constructor] interface x { };
};
#errors
2;xattr not applicable;w;Constructor
3;duplicate qname;m

#data
module m {
interface x {
  [IndexGetter] long method (in unsigned long x);
};
};
#errors
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method (in unsigned long x);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] void method (in unsigned long x);
};
};
#errors
#document
| module m {
| 
| interface x {
|   [IndexGetter] void method (in unsigned long x);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] long method (in long x);
};
};
#errors
3;wrong signature accessor;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method (in long x);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] long method (in DOMString x);
};
};
#errors
3;wrong signature accessor;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method (in DOMString x);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] long method (in unsigned long x, in DOMString y);
};
};
#errors
3;wrong signature accessor;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method (in unsigned long x, in DOMString y);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] long method ();
};
};
#errors
3;wrong signature accessor;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method ();
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] long method (in unsigned long x);
  [IndexGetter] boolean method2 (in unsigned long x);
};
};
#errors
4;duplicate accessor;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] long method (in unsigned long x);
|   [IndexGetter] boolean method2 (in unsigned long x);
| };
| 
| };

#data
module m {
interface x {
  [IndexGetter] attribute long m;
};
};
#errors
3;xattr not applicable;w;IndexGetter
#document
| module m {
| 
| interface x {
|   [IndexGetter] attribute long m;
| };
| 
| };

#data
module m {
  interface x {
    [IndexGetter] long m (in unsigned long n);
  };
  interface y {
    [IndexGetter] long m (in unsigned long n);
  };
};
#errors

#data
module m {
  interface x {
    [IndexGetter=a] long m (in unsigned long n);
  };
};
#errors
3;xattr id not allowed;m;IndexGetter

#data
module m {
  interface x {
    [IndexGetter()] long m (in unsigned long n);
  };
};
#errors
3;xattr arglist not allowed;m;IndexGetter

#data
module m {
  interface x {
    [IndexGetter=a()] long m (in unsigned long n);
  };
};
#errors
3;xattr id not allowed;m;IndexGetter
3;xattr arglist not allowed;m;IndexGetter

#data
module m {
interface x {
  [IndexGetter,IndexGetter] long method (in unsigned long x);
};
};
#errors
3;duplicate xattr;w;IndexGetter

#data
module m {
interface x {
  [IndexGetter,IndexSetter] long method (in unsigned long x);
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [NameGetter,IndexGetter] long method (in unsigned long x);
};
};
#errors
3;wrong signature accessor;w;NameGetter

#data
module m {
interface x {
  [IndexSetter] any y ();
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter] any y (in long a);
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter] any y (in unsigned long a);
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter] any y (in unsigned long a, in any b);
};
};
#errors

#data
module m {
interface x {
  [IndexSetter] void y (in unsigned long a, in any b);
};
};
#errors

#data
module m {
interface x {
  [IndexSetter] any y (in DOMString a, in any b);
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter] any y (in unsigned long a, in any b, in any c);
};
};
#errors
3;wrong signature accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter] any y (in unsigned long a, in any b);
  [IndexSetter] any z (in unsigned long a, in any b);
};
};
#errors
4;duplicate accessor;w;IndexSetter

#data
module m {
interface x {
  [IndexGetter] any y (in unsigned long a);
  [IndexSetter] any z (in unsigned long a, in DOMString b);
};
};
#errors

#data
module m {
interface x {
  [IndexSetter] attribute unsigned long y;
};
};
#errors
3;xattr not applicable;w;IndexSetter

#data
module m {
interface x {
  [IndexSetter=x] any y (in unsigned long a, in any b);
};
};
#errors
3;xattr id not allowed;m;IndexSetter

#data
module m {
interface x {
  [IndexSetter()] any y (in unsigned long a, in any b);
};
};
#errors
3;xattr arglist not allowed;m;IndexSetter

#data
module m {
interface x {
  [IndexSetter=x()] any y (in unsigned long a, in any b);
};
};
#errors
3;xattr id not allowed;m;IndexSetter
3;xattr arglist not allowed;m;IndexSetter

#data
module m {
interface x {
  [IndexSetter,IndexSetter] long method (in unsigned long x,
      in unsigned short y);
};
};
#errors
3;duplicate xattr;w;IndexSetter

#data
module m {
interface x {
  [NameGetter] long method (in DOMString x);
};
};
#errors
#document
| module m {
| 
| interface x {
|   [NameGetter] long method (in DOMString x);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] void method (in DOMString x);
};
};
#errors
#document
| module m {
| 
| interface x {
|   [NameGetter] void method (in DOMString x);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] long method (in long x);
};
};
#errors
3;wrong signature accessor;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] long method (in long x);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] long method (in unsigned long x);
};
};
#errors
3;wrong signature accessor;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] long method (in unsigned long x);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] long method (in DOMString x, in DOMString y);
};
};
#errors
3;wrong signature accessor;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] long method (in DOMString x, in DOMString y);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] long method ();
};
};
#errors
3;wrong signature accessor;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] long method ();
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] long method (in DOMString x);
  [NameGetter] boolean method2 (in DOMString x);
};
};
#errors
4;duplicate accessor;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] long method (in DOMString x);
|   [NameGetter] boolean method2 (in DOMString x);
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] attribute long m;
};
};
#errors
3;xattr not applicable;w;NameGetter
#document
| module m {
| 
| interface x {
|   [NameGetter] attribute long m;
| };
| 
| };

#data
module m {
interface x {
  [NameGetter] attribute DOMString m;
};
};
#errors
3;xattr not applicable;w;NameGetter

#data
module m {
interface x {
  [NameGetter] attribute _DOMString m;
};
};
#errors
3;xattr not applicable;w;NameGetter
3;type not defined;m;DOMString

#data
module m {
  interface x {
    [NameGetter] long m (in DOMString n);
  };
  interface y {
    [NameGetter] long m (in DOMString n);
  };
};
#errors

#data
module m {
  interface x {
    [NameGetter=a] long m (in DOMString n);
  };
};
#errors
3;xattr id not allowed;m;NameGetter

#data
module m {
  interface x {
    [NameGetter()] long m (in DOMString n);
  };
};
#errors
3;xattr arglist not allowed;m;NameGetter

#data
module m {
  interface x {
    [NameGetter=a()] long m (in DOMString n);
  };
};
#errors
3;xattr id not allowed;m;NameGetter
3;xattr arglist not allowed;m;NameGetter

#data
module m {
interface x {
  [NameGetter,NameGetter] long method (in DOMString x);
};
};
#errors
3;duplicate xattr;w;NameGetter

#data
module m {
interface x {
  [NameSetter] any y ();
};
};
#errors
3;wrong signature accessor;w;NameSetter

#data
module m {
interface x {
  [NameSetter] any y (in DOMString a);
};
};
#errors
3;wrong signature accessor;w;NameSetter

#data
module m {
interface x {
  [NameSetter] any y (in unsigned long a);
};
};
#errors
3;wrong signature accessor;w;NameSetter

#data
module m {
interface x {
  [NameSetter] any y (in DOMString a, in any b);
};
};
#errors

#data
module m {
interface x {
  [NameSetter] void y (in DOMString a, in any b);
};
};
#errors

#data
module m {
interface x {
  [NameSetter] any y (in unsigned long a, in any b);
};
};
#errors
3;wrong signature accessor;w;NameSetter

#data
module m {
interface x {
  [NameSetter] any y (in DOMString a, in any b, in any c);
};
};
#errors
3;wrong signature accessor;w;NameSetter

#data
module m {
interface x {
  [NameSetter] any y (in _DOMString a, in any b);
};
};
#errors
3;wrong signature accessor;w;NameSetter
3;type not defined;m;DOMString

#data
module m {
interface x {
  [NameSetter] any y (in DOMString a, in any b);
  [NameSetter] any z (in DOMString a, in any b);
};
};
#errors
4;duplicate accessor;w;NameSetter

#data
module m {
interface x {
  [NameGetter] any y (in DOMString a);
  [NameSetter] any z (in DOMString a, in DOMString b);
};
};
#errors

#data
module m {
interface x {
  [IndexSetter] any y (in unsigned long a, in DOMString b);
  [NameSetter] any z (in DOMString a, in DOMString b);
};
};
#errors

#data
module m {
interface x {
  [NameSetter] attribute DOMString y;
};
};
#errors
3;xattr not applicable;w;NameSetter

#data
module m {
interface x {
  [NameSetter=x] any y (in DOMString a, in any b);
};
};
#errors
3;xattr id not allowed;m;NameSetter

#data
module m {
interface x {
  [NameSetter()] any y (in DOMString a, in any b);
};
};
#errors
3;xattr arglist not allowed;m;NameSetter

#data
module m {
interface x {
  [NameSetter=x()] any y (in DOMString a, in any b);
};
};
#errors
3;xattr id not allowed;m;NameSetter
3;xattr arglist not allowed;m;NameSetter

#data
module m {
interface x {
  [NameSetter,NameSetter] long method (in DOMString y,in unsigned long x);
};
};
#errors
3;duplicate xattr;w;NameSetter

#data
module m {
interface x {
  [Null=Null] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Null=Null] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Null=Empty] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Null=Empty] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Null=Null,Null=Empty] attribute DOMString x;
};
};
#errors
3;duplicate xattr;w;Null

#data
module m {
interface x {
  void m ([Null=Null,Null=Empty] in DOMString x);
};
};
#errors
3;duplicate xattr;w;Null

#data
module m {
interface x {
  [Null=null] attribute DOMString x;
};
};
#errors
3:null;xattr id value not allowed;m;Null

#data
module m {
interface x {
  void m ([Null=null] in DOMString x);
};
};
#errors
3:null;xattr id value not allowed;m;Null

#data
module m {
interface x {
  [Null] attribute DOMString x;
};
};
#errors
3;xattr id missing;m;Null

#data
module m {
interface x {
  void m ([Null] in DOMString x);
};
};
#errors
3;xattr id missing;m;Null

#data
module m {
interface x {
  [Null()] attribute DOMString x;
};
};
#errors
3;xattr id missing;m;Null
3;xattr arglist not allowed;m;Null

#data
module m {
interface x {
  void m ([Null()] in DOMString x);
};
};
#errors
3;xattr id missing;m;Null
3;xattr arglist not allowed;m;Null

#data
module m {
interface x {
  [Null=Null()] attribute DOMString x;
};
};
#errors
3;xattr arglist not allowed;m;Null

#data
module m {
interface x {
  void m ([Null=Null()] in DOMString x);
};
};
#errors
3;xattr arglist not allowed;m;Null

#data
module m {
interface x {
  [Null=Null] attribute _DOMString x;
};
};
#errors
3;xattr for wrong type;w;Null
3;type not defined;m;DOMString

#data
module m {
interface x {
  void m ([Null=Null] in _DOMString x);
};
};
#errors
3;xattr for wrong type;w;Null
3;type not defined;m;DOMString

#data
module m {
interface x {
  [Null=_Null] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Null=_Null] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Null=_Empty] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Null=_Empty] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  void m ([Null=Undefined] in DOMString x);
};
};
#errors
3:Undefined;xattr id value not allowed;m;Null

#data
module m {
[Null=Null]interface x {
  attribute DOMString x;
};
};
#errors
2;xattr not applicable;w;Null

#data
module m {
interface x {
  [Null=Empty] void m (in DOMString x);
};
};
#errors
3;xattr not applicable;w;Null

#data
module m {
interface x {
  [null=Null] attribute DOMString x;
};
};
#errors
3;unknown xattr;u;null

#data
module m {
interface x {
  void m ([null=Null] in DOMString x);
};
};
#errors
3;unknown xattr;u;null

#data
module m {
interface y {
  attribute DOMString p;
};
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors

#data
module m {
interface y {
};
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors
5:p;referenced attr not defined;m;PutForwards

#data
module m {
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors
3;type not defined;m;y
3;attr type not interface;w;PutForwards

#data
module m {
interface x {
  [PutForwards=p] readonly attribute DOMString n;
};
};
#errors
3;attr type not interface;w;PutForwards

#data
module m {
exception y {
  long p;
};
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors
6;type not defined;m;y
6;attr type not interface;w;PutForwards

#data
module m {
interface y {
  attribute DOMString p;
};
interface x {
  [PutForwards=p] attribute y n;
};
};
#errors
6;attr not readonly;w;PutForwards

#data
module m {
interface y {
  readonly attribute DOMString p;
};
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors

#data
module m {
interface y {
  attribute DOMString p;
};
interface x {
  [PutForwards=p, Null=Empty] readonly attribute y n;
};
};
#errors
6;xattr for wrong type;w;Null

#data
module m {
interface y {
  [Null=Empty] attribute DOMString p;
};
interface x {
  [PutForwards=p, Null=Null] readonly attribute y n;
};
};
#errors
6;xattr for wrong type;w;Null

#data
module m {
typedef sequence<unsigned short> y;
interface x {
  [PutForwards=p] readonly attribute y n;
};
};
#errors
4;attr type not interface;w;PutForwards

#data
module m {
interface x {
  [PutForwards=p] readonly attribute y n;
};
interface y {
  attribute DOMString p;
};
};
#errors
3;type not defined;m;y
3;attr type not interface;w;PutForwards

#data
module m {
interface y;
interface x {
  [PutForwards=p] readonly attribute y n;
};
interface y {
  attribute DOMString p;
};
};
#errors
6;duplicate qname;m

#data
module m {
interface y {
  attribute DOMString p;
};
interface x {
  [PutForwards=p,PutForwards=p] readonly attribute y n;
};
};
#errors
6;duplicate xattr;w;PutForwards

#data
module m {
interface y {
  attribute DOMString p;
  attribute long q;
};
interface x {
  [PutForwards=p,PutForwards=q] readonly attribute y n;
};
};
#errors
7;duplicate xattr;w;PutForwards

#data
module m {
[Stringifies=a,Stringifies=b] interface x {
  attribute boolean a;
  attribute long b;
};
};
#errors
2;duplicate xattr;w;Stringifies

#data
module m {
[Stringifies] exception x {
};
};
#errors
2;xattr not applicable;w;Stringifies

#data
[Stringifies] module m {
interface x {
};
};
#errors
1;xattr not applicable;w;Stringifies

#data
module m {
interface x {
  [Stringifies] attribute DOMString b;
};
};
#errors
3;xattr not applicable;w;Stringifies

#data
module m {
[Stringifies()] interface x {
};
};
#errors
2;xattr arglist not allowed;m;Stringifies

#data
module m {
[Stringifies=a()] interface x {
};
};
#errors
2;xattr arglist not allowed;m;Stringifies
2:a;referenced attr not defined;m;Stringifies

#data
module m {
interface x {
  void m ([Variadic] in long a);
};
};
#errors

#data
module m {
interface x {
  void m (in long b, [Variadic] in long a);
};
};
#errors

#data
module m {
interface x {
  void m ([Variadic] in long a, in DOMString b);
};
};
#errors
3;argument after variadic;w

#data
module m {
interface x {
  void m ([Variadic] in long a);
  void n (in long b);
};
};
#errors

#data
module m {
interface x {
  void m ([Variadic=a] in long a);
};
};
#errors
3;xattr id not allowed;m;Variadic

#data
module m {
interface x {
  void m ([Variadic()] in long a);
};
};
#errors
3;xattr arglist not allowed;m;Variadic

#data
module m {
interface x {
  void m ([Variadic=a()] in long a);
};
};
#errors
3;xattr id not allowed;m;Variadic
3;xattr arglist not allowed;m;Variadic

#data
module m {
interface x {
  void m (in long b, [Variadic,Variadic] in long a);
};
};
#errors
3;duplicate xattr;w;Variadic

#data
module m {
[NamedConstructor=a]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a()]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a(in long b)]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a,NamedConstructor=b]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a,NamedConstructor=a]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a,NamedConstructor=a()]
interface x {};
};
#errors

#data
module m {
[NamedConstructor=a]
interface x {};
[NamedConstructor=a]
interface y {};
};
#errors
4:a;duplicate constructor name;m

#data
module m {
[NamedConstructor=a]
interface x {};
};
[NamedConstructor=a]
interface y{};
#errors
6;non-module definition;s
5:a;duplicate constructor name;m

#data
module m {
[NamedConstructor=a]
interface x {};
[NamedConstructor=_a]
interface y {};
};
#errors
4:a;duplicate constructor name;m

#data
module m {
[NamedConstructor=a]
interface x {};
};
module n {
[NamedConstructor=a]
interface y {};
};
#errors
6:a;duplicate constructor name;m

#data
module m {
[NamedConstructor=x]
interface x {};
};
#errors
2:x;duplicate constructor name;m

#data
module m {
[NamedConstructor=y]
interface x {};
interface y {};
};
#errors
2:y;duplicate constructor name;m

#data
module m {
exception a {};
[NamedConstructor=a]
interface x {};
};
#errors
3:a;duplicate constructor name;m

#data
module m {
[NamedConstructor=a]
interface x {};
};
module n {
  interface a {};
};
#errors
2:a;duplicate constructor name;m

#data
module m {
[NamedConstructor]
interface x {};
};
#errors
2;xattr id missing;m;NamedConstructor

#data
module m {
[NamedConstructor()]
interface x {};
};
#errors
2;xattr id missing;m;NamedConstructor

#data
module m {
[NamedConstructor=a]
exception x {};
};
#errors

#data
module m {
interface x {
  [NamedConstructor=a] void m ();
};
exception a {};
};
#errors
3;xattr not applicable;w;NamedConstructor

#data
module m {
[NamedConstructor=a]
interface x;
interface x {};
};
#errors
2;xattr not applicable;w;NamedConstructor
4;duplicate qname;m

#data
module m {
[NativeObject] interface x {};
};
#errors

#data
module m {
[NativeObject=FunctionOnly] interface x {};
};
#errors

#data
module m {
[NativeObject=functiononly] interface x {};
};
#errors
2:functiononly;xattr id value not allowed;m;NativeObject

#data
module m {
[NativeObject()] interface x {};
};
#errors
2;xattr arglist not allowed;m;NativeObject

#data
module m {
[NativeObject=FunctionOnly()] interface x {};
};
#errors
2;xattr arglist not allowed;m;NativeObject

#data
module m {
[NativeObject] exception x {};
};
#errors
2;xattr not applicable;w;NativeObject

#data
module m {
[NativeObject] interface x;
interface x {};
};
#errors
2;xattr not applicable;w;NativeObject
3;duplicate qname;m

#data
module m {
[NativeObject,NativeObject] interface x {};
};
#errors
2;duplicate xattr;w;NativeObject

#data
module m {
[NativeObject,NativeObject=FunctionOnly] interface x {};
};
#errors
2;duplicate xattr;w;NativeObject

#data
module m {
[NoInterfaceObject] interface x {};
};
#errors

#data
module m {
[NoInterfaceObject,NoInterfaceObject] interface x {};
};
#errors
2;duplicate xattr;w;NoInterfaceObject

#data
module m {
[NoInterfaceObject=a] interface x {};
};
#errors
2;xattr id not allowed;m;NoInterfaceObject

#data
module m {
[NoInterfaceObject()] interface x {};
};
#errors
2;xattr arglist not allowed;m;NoInterfaceObject

#data
module m {
[NoInterfaceObject=a()] interface x {};
};
#errors
2;xattr id not allowed;m;NoInterfaceObject
2;xattr arglist not allowed;m;NoInterfaceObject

#data
module m {
[NoInterfaceObject] exception x {};
};
#errors

#data
[NoInterfaceObject] module m {
interface x {};
};
#errors
1;xattr not applicable;w;NoInterfaceObject

#data
module m {
[NoInterfaceObject] interface x;
interface x {};
};
#errors
2;xattr not applicable;w;NoInterfaceObject
3;duplicate qname;m

#data
module m {
[NoInterfaceObject,NamedConstructor=a] interface x {};
};
#errors

#data
module m {
[NamedConstructor=x] interface y {};
[NoInterfaceObject] interface x {};
};
#errors

#data
module m {
[NoInterfaceObject,NamedConstructor=x] interface x {};
};
#errors

#data
module m {
[NamedConstructor=x] interface y {};
[NoInterfaceObject] exception x {};
};
#errors

#data
module m {
interface x {
  [Undefined=Null] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Undefined=Null] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Undefined=Empty] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Undefined=Empty] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Undefined=Null,Undefined=Empty] attribute DOMString x;
};
};
#errors
3;duplicate xattr;w;Undefined

#data
module m {
interface x {
  void m ([Undefined=Null,Undefined=Empty] in DOMString x);
};
};
#errors
3;duplicate xattr;w;Undefined

#data
module m {
interface x {
  void m ([Undefined=Null,Null=Empty] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Undefined=null] attribute DOMString x;
};
};
#errors
3:null;xattr id value not allowed;m;Undefined

#data
module m {
interface x {
  void m ([Undefined=null] in DOMString x);
};
};
#errors
3:null;xattr id value not allowed;m;Undefined

#data
module m {
interface x {
  [Undefined] attribute DOMString x;
};
};
#errors
3;xattr id missing;m;Undefined

#data
module m {
interface x {
  void m ([Undefined] in DOMString x);
};
};
#errors
3;xattr id missing;m;Undefined

#data
module m {
interface x {
  [Undefined()] attribute DOMString x;
};
};
#errors
3;xattr id missing;m;Undefined
3;xattr arglist not allowed;m;Undefined

#data
module m {
interface x {
  void m ([Undefined()] in DOMString x);
};
};
#errors
3;xattr id missing;m;Undefined
3;xattr arglist not allowed;m;Undefined

#data
module m {
interface x {
  [Undefined=Null()] attribute DOMString x;
};
};
#errors
3;xattr arglist not allowed;m;Undefined

#data
module m {
interface x {
  void m ([Undefined=Null()] in DOMString x);
};
};
#errors
3;xattr arglist not allowed;m;Undefined

#data
module m {
interface x {
  [Undefined=Null] attribute _DOMString x;
};
};
#errors
3;xattr for wrong type;w;Undefined
3;type not defined;m;DOMString

#data
module m {
interface x {
  void m ([Undefined=Null] in _DOMString x);
};
};
#errors
3;xattr for wrong type;w;Undefined
3;type not defined;m;DOMString

#data
module m {
interface x {
  [Undefined=_Null] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Undefined=_Null] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  [Undefined=_Empty] attribute DOMString x;
};
};
#errors

#data
module m {
interface x {
  void m ([Undefined=_Empty] in DOMString x);
};
};
#errors

#data
module m {
interface x {
  void m ([Undefined=Undefined] in DOMString x);
};
};
#errors
3:Undefined;xattr id value not allowed;m;Undefined

#data
module m {
[Undefined=Null]interface x {
  attribute DOMString x;
};
};
#errors
2;xattr not applicable;w;Undefined

#data
module m {
interface x {
  [Undefined=Empty] void m (in DOMString x);
};
};
#errors
3;xattr not applicable;w;Undefined
